LayoutTests/inspector/console/x-frame-options-message-expected.txt

CONSOLE MESSAGE: line 41: The X-Frame-Option 'deny' supplied in a <meta> element was ignored. X-Frame-Options may only be provided by an HTTP header sent with the document.


== Running test suite: Console.XFrameOptionsMessages
-- Running test case: XFrameOptionsDeny
Evaluating expression: triggerXFrameOptionDeny();
PASS: ConsoleMessage type should be 'security'.
PASS: ConsoleMessage level should be 'error'.
